retry: implement exponential backoff for event retry
To complement circuit breaker, we now have an exponential backoff retry
strategy. The common increasing bound strategy is implemented.

A small bug was exposed in the implementation that blocked the retry
forever if the strategy did not probe. Now, we simply wait for the retry
and proceed. This may change behavior in the breaker case, but is more
correct.

Signed-off-by: Stephen J Day <stephen.day@docker.com>